refactor: move captureCount StateFlow and camera hook registration to correct layers
- CameraManager exposes registerStateCallback/unregisterStateCallback
  instead of owning a StateFlow, consistent with FCManager and SimManager
- CameraHandler owns _captureCount MutableStateFlow and wasStoringPhoto,
  registers the DJI SystemState callback after initCameras succeeds,
  and unregisters in unload()
- CameraController registers handler.onImageCaptured in its init block,
  removing the camera-specific wiring from MAVLinkController.attachClient
